refactor(*): search and map pool refactoring
