+++ src-bellumgens/app/home/home.component.html
@@ -3,7 +3,7 @@
     <div class="slide slide1">
       <div class="slide-info">
         <h2><igx-icon fontSet="partners" name="eb-league-white" class="event-icon"></igx-icon>Esports Business League</h2>
-        <p>Esports Business League Season 2 finals <span class="highlight">7-8 November</span>!</p>
+        <p>Esports Business League Season 2 has concluded. Check out who the new champions are!</p>
         <a [href]="environment.ebleague" target="_blank" rel="noopener" igxButton="raised">Go to EB-League</a>
       </div>
     </div>
@@ -13,7 +13,7 @@
     <div class="slide slide0">
       <div class="slide-info">
         <h2><igx-icon fontSet="partners" name="bge-white" class="event-icon"></igx-icon>Bellum Gens Elite - Stara Zagora</h2>
-        <p>The first professional Bellum Gens CS:GO tournament was held in the city of Stara Zagora!</p>
+        <p>The first professional Bellum Gens CS:GO tournament was held in the city of Stara Zagora! Check out who the champions are!</p>
         <a href="https://bge-starazagora.com" target="_blank" rel="noopener" igxButton="raised">Go to BGE-Stara Zagora</a>
       </div>
     </div>
