+++ src-bellumgens/app/search/quick-search/quick-search.component.html
@@ -8,33 +8,33 @@
           </div>
         </div>
     </ng-template>
-    <igx-list-item *ngIf="searchResult.SteamUser?.steamUser" [isHeader]="true">Steam User</igx-list-item>
-    <igx-list-item *ngIf="searchResult.SteamUser?.steamUser"
-                  [routerLink]="['/players', searchResult.SteamUser.steamUser.customURL || searchResult.SteamUser.steamUser.steamID64]"
+    <igx-list-item *ngIf="searchResult.steamUser?.steamUser" [isHeader]="true">Steam User</igx-list-item>
+    <igx-list-item *ngIf="searchResult.steamUser?.steamUser"
+                  [routerLink]="['/players', searchResult.steamUser.steamUser.customURL || searchResult.steamUser.steamUser.steamID64]"
                   class="navigatable">
-      <igx-avatar igxListThumbnail [src]="searchResult.SteamUser.steamUser.avatarIcon" [initials]="searchResult.SteamUser.steamUser.realname" [roundShape]="true">
+      <igx-avatar igxListThumbnail [src]="searchResult.steamUser.steamUser.avatarIcon" [initials]="searchResult.steamUser.steamUser.realname" [roundShape]="true">
       </igx-avatar>
-      <igx-avatar *ngIf="searchResult.SteamUser.steamUser.country"
-                  [src]="searchResult.SteamUser.steamUser.country | playerCountry"
+      <igx-avatar *ngIf="searchResult.steamUser.steamUser.country"
+                  [src]="searchResult.steamUser.steamUser.country | playerCountry"
                   [roundShape]="true"
-                  [title]="searchResult.SteamUser.steamUser.country"
+                  [title]="searchResult.steamUser.steamUser.country"
                   class="country-badge-small">
       </igx-avatar>
 
       <h2 igxListLineTitle>
-        {{ searchResult.SteamUser.steamUser.steamID }}
+        {{ searchResult.steamUser.steamUser.steamID }}
       </h2>
-      <h4 igxListLineSubTitle>{{ searchResult.SteamUser.steamUser.realname }}</h4>
+      <h4 igxListLineSubTitle>{{ searchResult.steamUser.steamUser.realname }}</h4>
 
-      <div igxListAction *ngIf="!searchResult.SteamUser.userStatsException else private">
+      <div igxListAction *ngIf="!searchResult.steamUser.userStatsException else private">
           <div class="horizontal-center">
-            <igx-circular-bar class="icon-avatar" [value]="searchResult.SteamUser.userStats?.killDeathRatio" [max]="2" [animate]="true" [textVisibility]="false" title="KDR"></igx-circular-bar>
-            <span>{{ searchResult.SteamUser.userStats?.killDeathRatio }}</span>
+            <igx-circular-bar class="icon-avatar" [value]="searchResult.steamUser.userStats?.killDeathRatio" [max]="2" [animate]="true" [textVisibility]="false" title="KDR"></igx-circular-bar>
+            <span>{{ searchResult.steamUser.userStats?.killDeathRatio }}</span>
           </div>
 
           <div class="horizontal-center">
             <igx-icon fontSet="weapon-icons" name="headshot" class="icon-avatar"></igx-icon>
-            <span>{{ searchResult.SteamUser.userStats?.headshotPercentage | number:'1.2' }}%</span>
+            <span>{{ searchResult.steamUser.userStats?.headshotPercentage | number:'1.2' }}%</span>
           </div>
       </div>
 
@@ -43,8 +43,8 @@
       </ng-template>
     </igx-list-item>
     <igx-list-item [isHeader]="true">Players</igx-list-item>
-    <igx-list-item *ngIf="!searchResult.Players.length"><span igxListLine>No matching players...</span></igx-list-item>
-    <igx-list-item *ngFor="let user of searchResult.Players | reduceQuickSearchResult"
+    <igx-list-item *ngIf="!searchResult.players.length"><span igxListLine>No matching players...</span></igx-list-item>
+    <igx-list-item *ngFor="let user of searchResult.players | reduceQuickSearchResult"
                   [routerLink]="['/players', user.customURL || user.id]"
                   class="navigatable">
       <igx-avatar igxListThumbnail [src]="user.avatarIcon" [initials]="user.realname" [roundShape]="true">
@@ -77,7 +77,7 @@
         <span>Private</span>
       </ng-template>
     </igx-list-item>
-    <igx-list-item *ngIf="searchResult.Players.length > 3">
+    <igx-list-item *ngIf="searchResult.players.length > 3">
       <button igxButton="flat" igxRipple class="submit-button" [routerLink]="['/search/players', 'name=' + term]">View more players...</button>
     </igx-list-item>
   </igx-list>
@@ -91,19 +91,15 @@
       </div>
     </ng-template>
     <igx-list-item [isHeader]="true">Teams</igx-list-item>
-    <igx-list-item *ngIf="!searchResult.Teams.length">No matching teams...</igx-list-item>
-    <igx-list-item *ngFor="let team of searchResult.Teams | reduceQuickSearchResult" class="navigatable" [routerLink]="['/team', team.CustomUrl]">
+    <igx-list-item *ngIf="!searchResult.teams.length">No matching teams...</igx-list-item>
+    <igx-list-item *ngFor="let team of searchResult.teams | reduceQuickSearchResult" class="navigatable" [routerLink]="['/team', team.CustomUrl]">
       <igx-avatar igxListThumbnail [src]="team.TeamAvatar" [initials]="team.TeamName" [roundShape]="true">
       </igx-avatar>
       <h2 igxListLineTitle>
         {{ team.TeamName }}
       </h2>
-
-      <div igxListAction>
-        {{ team.Members.length }} Members
-      </div>
     </igx-list-item>
-    <igx-list-item *ngIf="searchResult.Teams.length > 3">
+    <igx-list-item *ngIf="searchResult.teams.length > 3">
       <button igxButton="flat" igxRipple class="submit-button" [routerLink]="['/search/teams', 'name=' + term]">View more teams...</button>
     </igx-list-item>
   </igx-list>
@@ -117,15 +113,15 @@
       </div>
     </ng-template>
     <igx-list-item [isHeader]="true">Strategies</igx-list-item>
-    <igx-list-item *ngIf="!searchResult.Strategies.length">No matching strategies...</igx-list-item>
-    <igx-list-item *ngFor="let strat of searchResult.Strategies | reduceQuickSearchResult" class="navigatable" [routerLink]="['/team', strat.CustomUrl]">
+    <igx-list-item *ngIf="!searchResult.strategies.length">No matching strategies...</igx-list-item>
+    <igx-list-item *ngFor="let strat of searchResult.strategies | reduceQuickSearchResult" class="navigatable" [routerLink]="['/team', strat.CustomUrl]">
       <igx-avatar igxListThumbnail [src]="strat.Image" [initials]="strat.title">
       </igx-avatar>
       <h2 igxListLineTitle>
         {{ strat.Title }}
       </h2>
     </igx-list-item>
-    <igx-list-item *ngIf="searchResult.Strategies.length > 3">
+    <igx-list-item *ngIf="searchResult.strategies.length > 3">
       <button igxButton="flat" igxRipple class="submit-button" [routerLink]="['/strategies', 'name=' + term]">View more strategies...</button>
     </igx-list-item>
   </igx-list>
