+++ src-bellumgens/app/search/search-results/player-results/player-results.component.html
@@ -22,7 +22,7 @@
         </div>
       </div>
 
-      <div class="horizontal-info__right" *ngIf="player.steamPrivate">
+      <div class="horizontal-center" *ngIf="player.steamPrivate">
         <span class="title">Private</span>
       </div>
     </igx-card-header>
@@ -30,19 +30,12 @@
       <div>Preferred roles</div>
       <igx-chips-area class="result-chips">
         <igx-chip [selectable]="false">
-          <span class='igx-chip__text'>{{ roles[player.primaryRole].name }}</span>
+          <span class="igx-chip__text">{{ roles[player.primaryRole].name }}</span>
         </igx-chip>
         <igx-chip [selectable]="false">
-          <span class='igx-chip__text'>{{ roles[player.secondaryRole].name }}</span>
+          <span class="igx-chip__text">{{ roles[player.secondaryRole].name }}</span>
         </igx-chip>
       </igx-chips-area>
-      <!-- <div>Available</div>
-      <igx-chips-area class="result-chips">
-        <igx-chip *ngFor="let weekday of player.availability | daysAvailable"
-                  [selectable]="false">
-          <span class='igx-chip__text'>{{ weekday.day | weekday }}</span>
-        </igx-chip>
-      </igx-chips-area> -->
     </igx-card-content>
   </igx-card>
 </div>
