+++ src-bellumgens/app/search/search-results/player-results/player-results.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { BellumgensApiService } from '../../../../../src-common/services/bellumgens-api.service';
 import { ALL_ROLES } from '../../../../../src-common/models/playerrole';
 import { BaseComponent } from '../../../base/base.component';
 import { IgxIconService } from '@infragistics/igniteui-angular';
 import { Title, Meta } from '@angular/platform-browser';
 import { CSGOPlayer } from '../../../../../src-common/models/csgoplayer';
+import { ApiSearchService } from '../../../../../src-common/services/bellumgens-api.search.service';
 
 @Component({
   selector: 'app-player-results',
@@ -18,8 +18,8 @@ export class PlayerResultsComponent extends BaseComponent {
   public roles = ALL_ROLES;
   public query: string;
 
-  constructor(private apiService: BellumgensApiService,
-              private iconService: IgxIconService,
+  constructor(private iconService: IgxIconService,
+              private apiService: ApiSearchService,
               title: Title,
               meta: Meta,
               activeRoute: ActivatedRoute) {
