+++ src-bellumgens/app/search/search-results/team-results/team-results.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { BellumgensApiService } from '../../../../../src-common/services/bellumgens-api.service';
 import { CSGOTeam } from '../../../../../src-common/models/csgoteam';
 import { BaseComponent } from '../../../base/base.component';
 import { Title, Meta } from '@angular/platform-browser';
+import { ApiSearchService } from '../../../../../src-common/services/bellumgens-api.search.service';
 
 @Component({
   selector: 'app-team-results',
@@ -15,7 +15,7 @@ export class TeamResultsComponent extends BaseComponent {
   public loading = false;
   public query: string;
 
-  constructor(private apiService: BellumgensApiService,
+  constructor(private apiService: ApiSearchService,
               title: Title,
               meta: Meta,
               activeRoute: ActivatedRoute) {
